Refactoring - extracted crac* files
