+++ src/hotspot/os/linux/os_linux.hpp
@@ -171,15 +171,6 @@ class Linux {
 
   static jlong fast_thread_cpu_time(clockid_t clockid);
 
-  static void vm_create_start();
-  static bool prepare_checkpoint();
-  static Handle checkpoint(jarray fd_arr, jobjectArray obj_arr, bool dry_run, jlong jcmd_stream, TRAPS);
-  static void restore();
-  static void close_extra_descriptors();
-
-  static jlong restore_start_time();
-  static jlong uptime_since_restore();
-
   // Determine if the vmid is the parent pid for a child in a PID namespace.
   // Return the namespace pid if so, otherwise -1.
   static int get_namespace_pid(int vmid);
