+++ src/hotspot/share/runtime/thread.cpp
@@ -2709,7 +2709,7 @@ jint Threads::check_for_restore(JavaVMInitArgs* args) {
     if (!Arguments::parse_options_for_restore(args)) {
       return JNI_ERR;
     }
-    os::Linux::restore();
+    crac::Linux::restore();
     if (!CRaCIgnoreRestoreIfUnavailable) {
       // FIXME switch to unified hotspot logging
       warning("cannot restore");
@@ -2785,7 +2785,7 @@ jint Threads::create_vm(JavaVMInitArgs* args, bool* canTryAgain) {
     os::pause();
   }
 
-  os::Linux::vm_create_start();
+  crac::Linux::vm_create_start();
 
   HOTSPOT_VM_INIT_BEGIN();
